+++ src/datasets/normalization/log_cp/script.py
@@ -4,11 +4,12 @@ import scanpy as sc
 par = {
     'input': "resources_test/common/pancreas/dataset.h5ad",
     'output': "output.h5ad",
-    'layer_output': "log_cpm",
-    'obs_size_factors': "log_cpm_size_factors"
+    'layer_output': "log_cp10k",
+    'obs_size_factors': "log_cp10k_size_factors",
+    'n_cp': 1e6,
 }
 meta = {
-    "functionality_name": "normalize_log_cpm"
+    "functionality_name": "normalize_log_cp10k"
 }
 ## VIASH END
 
@@ -18,7 +19,7 @@ adata = sc.read_h5ad(par['input'])
 print(">> Normalize data", flush=True)
 norm = sc.pp.normalize_total(
     adata, 
-    target_sum=1e6, 
+    target_sum=par["n_cp"], 
     layer="counts", 
     inplace=False
 )
@@ -27,7 +28,7 @@ lognorm = sc.pp.log1p(norm["X"])
 print(">> Store output in adata", flush=True)
 adata.layers[par["layer_output"]] = lognorm
 adata.obs[par["obs_size_factors"]] = norm["norm_factor"]
-adata.uns["normalization_id"] = meta["functionality_name"]
+adata.uns["normalization_id"] = par["norm_id"]
 
 print(">> Write data", flush=True)
 adata.write_h5ad(par['output'], compression="gzip")