+++ src/tasks/denoising/workflows/run/run_test_on_tower.sh
@@ -8,7 +8,7 @@ id: pancreas_subsample
 input_train: s3://openproblems-data/$DATASET_DIR/train.h5ad
 input_test: s3://openproblems-data/$DATASET_DIR/test.h5ad
 dataset_id: pancreas
-normalization_id: log_cpm
+normalization_id: log_cp10k
 output: scores.tsv
 publish_dir: s3://openproblems-nextflow/output_test/v2/denoising
 HERE