+++ src/tasks/dimensionality_reduction/metrics/distance_correlation/script.py
@@ -1,7 +1,7 @@
 import anndata as ad
 import numpy as np
 import sklearn.decomposition
-import scipy.optimize
+import scipy.stats
 import scipy.spatial
 from sklearn.metrics import pairwise_distances
 import umap
@@ -15,13 +15,13 @@ par = {
 }
 ## VIASH END
 
-def _rmse(X, X_emb):
+def _distance_correlation(X, X_emb):
     high_dimensional_distance_vector = scipy.spatial.distance.pdist(X)
     low_dimensional_distance_vector = scipy.spatial.distance.pdist(X_emb)
-    _, rmse = scipy.optimize.nnls(
-        low_dimensional_distance_vector[:, None], high_dimensional_distance_vector
+    corr = scipy.stats.spearmanr(
+        low_dimensional_distance_vector, high_dimensional_distance_vector
     )
-    return rmse
+    return corr
 
 print("Load data", flush=True)
 input_solution = ad.read_h5ad(par["input_solution"])
@@ -31,17 +31,18 @@ high_dim = input_solution.layers["normalized"]
 X_emb = input_embedding.obsm["X_emb"]
 
 print("Compute NNLS residual after SVD", flush=True)
-n_svd = 200
+n_svd = 500
 svd_emb = sklearn.decomposition.TruncatedSVD(n_svd).fit_transform(high_dim)
-rmse = _rmse(svd_emb, X_emb)
+dist_corr = _distance_correlation(svd_emb, X_emb)
 
+#! Explicitly not changing it to use diffusion map method as this will have a positive effect on the diffusion map method for this specific metric.
 print("Compute NLSS residual after spectral embedding", flush=True)
-n_comps = min(200, min(input_solution.shape) - 2)
+n_comps = min(1000, min(input_solution.shape) - 2)
 umap_graph = umap.UMAP(transform_mode="graph").fit_transform(high_dim)
 spectral_emb = umap.spectral.spectral_layout(
     high_dim, umap_graph, n_comps, random_state=np.random.default_rng()
 )
-rmse_spectral = _rmse(spectral_emb, X_emb)
+dist_corr_spectral = _distance_correlation(spectral_emb, X_emb)
 
 print("Create output AnnData object", flush=True)
 output = ad.AnnData(
@@ -49,8 +50,8 @@ output = ad.AnnData(
         "dataset_id": input_solution.uns["dataset_id"],
         "normalization_id": input_solution.uns["normalization_id"],
         "method_id": input_embedding.uns["method_id"],
-        "metric_ids": [ "rmse", "rmse_spectral" ],
-        "metric_values": [ rmse, rmse_spectral ]
+        "metric_ids": [ "distance correlation", "distance_correlation_spectral" ],
+        "metric_values": [ dist_corr, dist_corr_spectral ]
     }
 )