+++ src/tasks/dimensionality_reduction/workflows/run/run_test.sh
@@ -21,7 +21,7 @@ nextflow \
   -resume \
   --id pancreas \
   --dataset_id pancreas \
-  --normalization_id log_cpm \
+  --normalization_id log_cp10k \
   --input $DATASET_DIR/dataset.h5ad \
   --input_solution $DATASET_DIR/solution.h5ad \
   --output scores.tsv \