+++ src/tasks/label_projection/resources_test_scripts/pancreas.sh
@@ -9,7 +9,7 @@ REPO_ROOT=$(git rev-parse --show-toplevel)
 # ensure that the command below is run from the root of the repository
 cd "$REPO_ROOT"
 
-RAW_DATA=resources_test/common/pancreas/dataset.h5ad
+RAW_DATA=resources_test/common/pancreas/cp10k_dataset.h5ad
 DATASET_DIR=resources_test/label_projection/pancreas
 
 if [ ! -f $RAW_DATA ]; then
@@ -49,7 +49,7 @@ nextflow \
   -resume \
   --id pancreas \
   --dataset_id pancreas \
-  --normalization_id log_cpm \
+  --normalization_id log_cp10k \
   --input_train $DATASET_DIR/train.h5ad \
   --input_test $DATASET_DIR/test.h5ad \
   --input_solution $DATASET_DIR/solution.h5ad \